UnifiedHID Roccat Vulcan support
And selective device disabling support
